refacto: extraction des reducers de test dans un fichier séparé
