+++ dbmongo/js/test/helpers/reducers.d.ts
@@ -1,0 +1,2 @@
+export function reducer(array: any[], reduce: any): any
+export function invertedReducer(array: any[], reduce: any): any
